watermark_remover_project/watermark_remover/core/lama_wrapper.py
from PIL import Image
import os

class LaMaWrapper:
    def __init__(self, model_path: str = None):
        """
        Initializes the LaMaWrapper.
        For now, this is a placeholder for loading the actual LaMa model.

        Args:
            model_path (str, optional): Path to the pre-trained LaMa model.
                                         Defaults to None, implying potential download
                                         or default location.
        """
        # TODO: Implement actual LaMa model loading here.
        # This would involve:
        # 1. Downloading pre-trained model weights if not available.
        # 2. Initializing the PyTorch model architecture.
        # 3. Loading the weights into the model.
        print(f"LaMaWrapper initialized. (Placeholder - Model path: {model_path})")
        self.model = None # Placeholder for the loaded model

    def remove_watermark(self, image_path: str, output_path: str) -> str:
        """
        Removes a watermark from an image.
        Currently uses a placeholder (grayscale conversion) instead of LaMa.

        Args:
            image_path (str): Path to the input image with a watermark.
            output_path (str): Path to save the processed image.

        Returns:
            str: Path to the processed image.

        Raises:
            FileNotFoundError: If the input image_path does not exist.
        """
        if not os.path.exists(image_path):
            raise FileNotFoundError(f"Input image not found: {image_path}")

        try:
            image = Image.open(image_path)

            # TODO: Replace this placeholder with actual LaMa model inference.
            # The actual process would involve:
            # 1. Preprocessing the image to the format LaMa expects.
            # 2. Generating a mask for the watermark (this is a crucial step,
            #    LaMa is an inpainting model and needs a mask).
            #    For a PoC, a simple mask covering a predefined area could be used,
            #    or a more sophisticated approach if time allows.
            # 3. Running the image and mask through the LaMa model.
            # 4. Postprocessing the output.
            print(f"Processing image: {image_path} (Placeholder: converting to grayscale)")
            processed_image = image.convert('L')

            # Ensure output directory exists
            os.makedirs(os.path.dirname(output_path), exist_ok=True)
            
            processed_image.save(output_path)
            print(f"Processed image saved to: {output_path}")
            
            return output_path
        except Exception as e:
            print(f"Error processing image {image_path}: {e}")
            raise

if __name__ == '__main__':
    # Example Usage (for testing the placeholder logic)
    # Create dummy input directories and a sample image if they don't exist
    
    # Paths for the __main__ block, distinct from API paths
    # Relative to this file (watermark_remover_project/watermark_remover/core/lama_wrapper.py)
    script_dir = os.path.dirname(__file__)
    base_dir = os.path.abspath(os.path.join(script_dir, "../../../data/wrapper_test"))
    
    input_dir = os.path.join(base_dir, "input")
    output_dir = os.path.join(base_dir, "output")
    
    os.makedirs(input_dir, exist_ok=True)
    os.makedirs(output_dir, exist_ok=True)
    print(f"Wrapper test directories created/ensured: {input_dir}, {output_dir}")

    sample_image_path = os.path.join(input_dir, "sample_watermarked_image.png")
    sample_output_path = os.path.join(output_dir, "sample_processed_image.png")

    # Create a dummy image for testing
    try:
        img = Image.new('RGB', (100, 100), color = 'blue') # Changed color for clarity
        img.save(sample_image_path)
        print(f"Created dummy image: {sample_image_path}")
    except Exception as e:
        print(f"Could not create dummy image: {e}")


    if os.path.exists(sample_image_path):
        print("\n--- Example LaMaWrapper Usage (Placeholder) ---")
        try:
            # Initialize wrapper (no model path needed for placeholder)
            lama_wrapper = LaMaWrapper()
            
            # Process the image
            result_path = lama_wrapper.remove_watermark(sample_image_path, sample_output_path)
            print(f"Placeholder processing complete. Output at: {result_path}")

            # Verify output (optional)
            if os.path.exists(result_path):
                print("Output image file exists.")
            else:
                print("Error: Output image file was not created.")
        except FileNotFoundError as fnf_error:
            print(f"Error in example usage: {fnf_error}")
        except Exception as ex:
            print(f"An unexpected error occurred in example usage: {ex}")
        print("--- End Example ---")
    else:
        print(f"Skipping example usage as dummy input image could not be created or found at {sample_image_path}")

```
