watermark_remover_project/requirements.txt
fastapi
uvicorn[standard]
celery
Pillow
numpy
# torch
# torchvision
# Potential LaMa-Cleaner specific dependencies (add as discovered)
